backend/requirements.txt
flask
flask_sqlalchemy
flask_login
flask_wtf
pandas
openpyxl
python-dotenv
werkzeug
pytest
